Move API and record merging tests to separate files
