+++ specifyweb/businessrules/tests/accession.py
@@ -1,7 +1,7 @@
 from django.db.models import ProtectedError
 
 from specifyweb.specify import models
-from specifyweb.specify.api_tests import ApiTests
+from specifyweb.specify.tests.test_api import ApiTests
 from ..exceptions import BusinessRuleException
