+++ specifyweb/businessrules/tests/appraisal.py
@@ -2,7 +2,7 @@ import datetime
 from django.db.models import ProtectedError
 
 from specifyweb.specify import models
-from specifyweb.specify.api_tests import ApiTests
+from specifyweb.specify.tests.test_api import ApiTests
 from ..exceptions import BusinessRuleException
 
 class AppraisalTests(ApiTests):
