+++ specifyweb/businessrules/tests/borrowagent.py
@@ -1,7 +1,7 @@
 
 from unittest import skip
 from specifyweb.specify import models
-from specifyweb.specify.api_tests import ApiTests
+from specifyweb.specify.tests.test_api import ApiTests
 from ..exceptions import BusinessRuleException
 
 class BorrowAgentTests(ApiTests):
