+++ specifyweb/businessrules/tests/collectionobject.py
@@ -1,5 +1,5 @@
 from specifyweb.specify import models
-from specifyweb.specify.api_tests import ApiTests
+from specifyweb.specify.tests.test_api import ApiTests
 from ..exceptions import BusinessRuleException
 
 class CollectionObjectTests(ApiTests):
