+++ specifyweb/businessrules/tests/gift.py
@@ -1,6 +1,6 @@
 from django.db import IntegrityError
 from specifyweb.specify import models
-from specifyweb.specify.api_tests import ApiTests
+from specifyweb.specify.tests.test_api import ApiTests
 from ..exceptions import BusinessRuleException
 
 class GiftTests(ApiTests):
