+++ specifyweb/businessrules/tests/journal.py
@@ -1,6 +1,6 @@
 from django.db.models import ProtectedError
 from specifyweb.specify import models
-from specifyweb.specify.api_tests import ApiTests
+from specifyweb.specify.tests.test_api import ApiTests
 
 class JournalBusinessRuleTests(ApiTests):
     def test_referenceworks_block_delete(self):
