+++ specifyweb/businessrules/tests/storage.py
@@ -3,7 +3,7 @@ from unittest import skip
 from django.db.models import ProtectedError
 
 from specifyweb.specify import models
-from specifyweb.specify.api_tests import ApiTests
+from specifyweb.specify.tests.test_api import ApiTests
 
 class StorageTests(ApiTests):
     def setUp(self):
