+++ specifyweb/businessrules/tests/taxontreedefitem.py
@@ -1,5 +1,5 @@
 from django.db.models import ProtectedError
-from specifyweb.specify.api_tests import ApiTests
+from specifyweb.specify.tests.test_api import ApiTests
 from ..exceptions import TreeBusinessRuleException
 from specifyweb.specify import models
