+++ specifyweb/context/tests.py
@@ -4,7 +4,7 @@ from django.test import TestCase, Client
 from jsonschema import validate  # type: ignore
 from jsonschema.exceptions import ValidationError  # type: ignore
 
-from specifyweb.specify.api_tests import ApiTests
+from specifyweb.specify.tests.test_api import ApiTests
 from . import viewsets
