+++ specifyweb/workbench/upload/tests/testonetoone.py
@@ -1,11 +1,10 @@
-import json
 from jsonschema import validate # type: ignore
-from typing import List, Dict, Any, NamedTuple, Union
-from specifyweb.specify.api_tests import get_table
+from typing import Dict
+from specifyweb.specify.tests.test_api import get_table
 from .base import UploadTestsBase
-from ..upload_result import Uploaded, Matched, NullRecord, ParseFailures, FailedBusinessRule
-from ..upload import do_upload, do_upload_csv
-from ..upload_table import UploadTable, OneToOneTable, ScopedUploadTable, ScopedOneToOneTable
+from ..upload_result import Uploaded, Matched, NullRecord
+from ..upload import do_upload
+from ..upload_table import UploadTable, OneToOneTable
 from ..upload_plan_schema import schema, parse_plan
 
 class OneToOneTests(UploadTestsBase):
