+++ specifyweb/workbench/upload/tests/testunupload.py
@@ -1,7 +1,7 @@
 from specifyweb.specify import auditcodes
-from specifyweb.specify.api_tests import get_table
+from specifyweb.specify.tests.test_api import get_table
 from .base import UploadTestsBase
-from ..upload import do_upload, do_upload_csv, unupload_record
+from ..upload import do_upload, unupload_record
 from ..upload_table import UploadTable
 from ..treerecord import TreeRecord
 from ..tomany import ToManyRecord
